Added random generation and fitting of std-poly
