+++ mfit_animator/__init__.py
@@ -1,0 +1,1 @@
+from .mfit_animator import startup
